+++ src/main/scala/edu/ie3/simona/agent/grid/GridAgentBuilder.scala
@@ -12,6 +12,7 @@ import edu.ie3.datamodel.models.input.system._
 import edu.ie3.simona.actor.SimonaActorNaming._
 import edu.ie3.simona.agent.EnvironmentRefs
 import edu.ie3.simona.agent.em.EmAgent
+import edu.ie3.simona.agent.grid.GridAgentData.GridAgentConstantData
 import edu.ie3.simona.agent.participant.ParticipantAgentInit.{
   ParticipantRefs,
   SimulationParameters,
@@ -47,70 +48,47 @@ import java.util.UUID
 import scala.jdk.CollectionConverters._
 import scala.jdk.OptionConverters.RichOptional
 
-/** Holds all methods that should be available to a [[GridAgent]]
+/** Provides functionality for building system participants of a [[GridAgent]].
   *
-  * @param gridAgentContext
-  *   ActorContext of the grid agent
-  * @param environmentRefs
-  *   The environment refs
-  * @param simulationStartDate
-  *   Simulation start date
-  * @param simulationEndDate
-  *   Simulation end date
-  * @param participantsConfig
-  *   Configuration information for participant models
-  * @param outputConfig
-  *   Configuration information for output behaviour
-  * @param resolution
-  *   The agents time resolution it wants to be triggered
-  * @param listener
-  *   System participant listeners
-  * @param log
-  *   The logging adapter to use here
   * @since 2019-07-18
   */
-class GridAgentBuilder(
-    gridAgentContext: ActorContext[GridAgent.Request],
-    environmentRefs: EnvironmentRefs,
-    simulationStartDate: ZonedDateTime,
-    simulationEndDate: ZonedDateTime,
-    emConfigs: AssetConfigs[EmRuntimeConfig],
-    participantsConfig: Participant,
-    outputConfig: AssetConfigs[ParticipantOutputConfig],
-    resolution: Long,
-    listener: Iterable[ActorRef[ResultEvent]],
-    log: Logger,
-) {
+object GridAgentBuilder {
+
+  /** Builds the relevant system participant for the given subgrid.
+    *
+    * @param subGridContainer
+    *   The subgrid container containing the system participants.
+    * @param thermalIslandGridsByBusId
+    *   Thermal islands by bus UUID.
+    * @param gridAgentContext
+    *   ActorContext of the grid agent.
+    * @param log
+    *   The logging adapter to use here.
+    */
   def buildSystemParticipants(
       subGridContainer: SubGridContainer,
       thermalIslandGridsByBusId: Map[UUID, ThermalGrid],
+  )(implicit
+      gridConf: GridAgentConstantData,
+      gridAgentContext: ActorContext[GridAgent.Request],
+      log: Logger,
   ): Map[UUID, Set[ActorRef[ParticipantAgent.Request]]] = {
 
     val systemParticipants =
-      filterSysParts(subGridContainer, environmentRefs)
-
-    val outputConfigUtil =
-      ConfigUtil.OutputConfigUtil.participants(outputConfig)
+      filterSysParts(subGridContainer, gridConf.environmentRefs)
 
     // ems that control at least one participant directly
     val firstLevelEms = systemParticipants.flatMap {
       _.getControllingEm.toScala.map(em => em.getUuid -> em)
     }.toMap
 
-    val allEms = buildEmsRecursively(
-      EmConfigUtil(emConfigs),
-      outputConfigUtil,
-      firstLevelEms,
-    )
+    val allEms = buildEmsRecursively(firstLevelEms)
 
     /* Browse through all system participants, build actors and map their node's UUID to the actor references */
     buildParticipantToActorRef(
-      participantsConfig,
       allEms,
-      outputConfigUtil,
       systemParticipants,
       thermalIslandGridsByBusId,
-      environmentRefs,
     )
   }
 
@@ -123,21 +101,21 @@ class GridAgentBuilder(
     * below.
     *
     * @param subGridContainer
-    *   the original subGrid container
+    *   The original subGrid container.
     * @return
-    *   a filtered subGrid container w/o assets for which no agent
-    *   implementations exist atm
+    *   A filtered subGrid container w/o assets for which no agent
+    *   implementations exist atm.
     */
   private def filterSysParts(
       subGridContainer: SubGridContainer,
       environmentRefs: EnvironmentRefs,
-  ) = {
+  )(implicit log: Logger): Seq[SystemParticipantInput] = {
 
     val (notProcessedElements, availableSysParts) =
       subGridContainer.getSystemParticipants
         .allEntitiesAsList()
         .asScala
-        .foldLeft((Set.empty[String], Vector.empty[SystemParticipantInput])) {
+        .foldLeft((Set.empty[String], Seq.empty[SystemParticipantInput])) {
           case (
                 (notProcessedElements, availableSystemParticipants),
                 curSysPart,
@@ -175,34 +153,23 @@ class GridAgentBuilder(
     * agents are also introduced to the agent environment and the scheduler is
     * requested to send an initialisation trigger.
     *
-    * @param participantsConfig
-    *   Configuration information for participant models
     * @param emAgents
-    *   mapping: em uuid to agent
-    * @param outputConfigUtil
-    *   Containing configuration information for output behaviour
+    *   Mapping: em uuid to agent.
     * @param participants
-    *   Set of system participants to create agents for
+    *   Set of system participants to create agents for.
     * @param thermalIslandGridsByBusId
-    *   Collection of thermal island grids, mapped by their thermal bus uuid
-    * @param environmentRefs
-    *   References to singleton entities representing the agent environment
+    *   Collection of thermal island grids, mapped by their thermal bus uuid.
     * @return
-    *   A map from coupling point to set of actor references
+    *   A map from coupling point to set of actor references.
     */
   private def buildParticipantToActorRef(
-      participantsConfig: Participant,
       emAgents: Map[UUID, ActorRef[FlexResponse]],
-      outputConfigUtil: OutputConfigUtil,
-      participants: Vector[SystemParticipantInput],
+      participants: Seq[SystemParticipantInput],
       thermalIslandGridsByBusId: Map[UUID, ThermalGrid],
-      environmentRefs: EnvironmentRefs,
+  )(implicit
+      gridConf: GridAgentConstantData,
+      gridAgentContext: ActorContext[GridAgent.Request],
   ): Map[UUID, Set[ActorRef[ParticipantAgent.Request]]] = {
-    /* Prepare the config util for the participant models, which (possibly) utilizes as map to speed up the initialization
-     * phase */
-    val participantConfigUtil =
-      ConfigUtil.ParticipantConfigUtil(participantsConfig)
-
     participants
       .map { participant =>
         val node = participant.getNode
@@ -220,12 +187,8 @@ class GridAgentBuilder(
             )
 
         val actorRef = buildParticipantActor(
-          participantsConfig.requestVoltageDeviationThreshold,
-          participantConfigUtil,
-          outputConfigUtil,
-          participant,
           thermalIslandGridsByBusId,
-          environmentRefs,
+          participant,
           controllingEm,
         )
         // return uuid to actorRef
@@ -240,23 +203,20 @@ class GridAgentBuilder(
     * its way up to EMs at root level, which are not EM-controlled themselves.
     * The first level can also be root level.
     *
-    * @param emConfigUtil
-    *   Configuration util for participant models
-    * @param outputConfigUtil
-    *   Configuration util for output behaviour
     * @param emInputs
     *   EMs of the current level, which can be controlled by further EMs at
-    *   higher levels
+    *   higher levels.
     * @param previousLevelEms
-    *   EMs that have been built by the previous recursion level
+    *   EMs that have been built by the previous recursion level.
     * @return
-    *   Map from model UUID to EmAgent ActorRef
+    *   Map from model UUID to EmAgent ActorRef.
     */
   private def buildEmsRecursively(
-      emConfigUtil: ConfigUtil.EmConfigUtil,
-      outputConfigUtil: OutputConfigUtil,
       emInputs: Map[UUID, EmInput],
       previousLevelEms: Map[UUID, ActorRef[FlexResponse]] = Map.empty,
+  )(implicit
+      gridConf: GridAgentConstantData,
+      gridAgentContext: ActorContext[GridAgent.Request],
   ): Map[UUID, ActorRef[FlexResponse]] = {
     // For the current level, split controlled and uncontrolled EMs.
     // Uncontrolled EMs can be built right away.
@@ -267,8 +227,6 @@ class GridAgentBuilder(
         else {
           val actor = buildEm(
             emInput,
-            emConfigUtil.getOrDefault(uuid),
-            outputConfigUtil.getOrDefault(NotifierIdentifier.Em),
             maybeControllingEm = None,
           )
           Right(uuid -> actor)
@@ -288,8 +246,6 @@ class GridAgentBuilder(
 
       // Return value includes previous level and uncontrolled EMs of this level
       val recursiveEms = buildEmsRecursively(
-        emConfigUtil,
-        outputConfigUtil,
         controllingEms,
         previousLevelAndUncontrolledEms,
       )
@@ -308,8 +264,6 @@ class GridAgentBuilder(
 
         uuid -> buildEm(
           emInput,
-          emConfigUtil.getOrDefault(uuid),
-          outputConfigUtil.getOrDefault(NotifierIdentifier.Em),
           maybeControllingEm = controllingEm,
         )
       }.toMap
@@ -321,108 +275,113 @@ class GridAgentBuilder(
   }
 
   private def buildParticipantActor(
-      requestVoltageDeviationThreshold: Double,
-      participantConfigUtil: ConfigUtil.ParticipantConfigUtil,
-      outputConfigUtil: OutputConfigUtil,
-      participantInputModel: SystemParticipantInput,
       thermalIslandGridsByBusId: Map[UUID, ThermalGrid],
-      environmentRefs: EnvironmentRefs,
+      participantInputModel: SystemParticipantInput,
       maybeControllingEm: Option[ActorRef[FlexResponse]],
+  )(implicit
+      gridConf: GridAgentConstantData,
+      gridAgentContext: ActorContext[GridAgent.Request],
   ): ActorRef[ParticipantAgent.Request] = {
 
     val serviceMap: Map[ServiceType, ActorRef[_ >: ServiceMessage]] =
       Seq(
-        Some(ServiceType.WeatherService -> environmentRefs.weather),
-        environmentRefs.evDataService.map(ref =>
+        Some(ServiceType.WeatherService -> gridConf.environmentRefs.weather),
+        gridConf.environmentRefs.evDataService.map(ref =>
           ServiceType.EvMovementService -> ref
         ),
       ).flatten.toMap
 
     val participantRefs = ParticipantRefs(
       gridAgentContext.self,
-      environmentRefs.primaryServiceProxy,
+      gridConf.environmentRefs.primaryServiceProxy,
       serviceMap,
-      listener,
+      gridConf.listener,
     )
 
     val simParams = SimulationParameters(
-      resolution,
-      Each(requestVoltageDeviationThreshold),
-      simulationStartDate,
-      simulationEndDate,
+      gridConf.resolution,
+      Each(
+        gridConf.simonaConfig.simona.runtime.participant.requestVoltageDeviationThreshold
+      ),
+      gridConf.simStartTime,
+      gridConf.simEndTime,
     )
 
     participantInputModel match {
       case input: FixedFeedInInput =>
         buildParticipant(
           SimpleInputContainer(input),
-          participantConfigUtil.getOrDefault[FixedFeedInRuntimeConfig](
+          gridConf.participantConfigUtil.getOrDefault[FixedFeedInRuntimeConfig](
             input.getUuid
           ),
-          outputConfigUtil.getOrDefault(NotifierIdentifier.FixedFeedIn),
+          gridConf.outputConfigUtil.getOrDefault(
+            NotifierIdentifier.FixedFeedIn
+          ),
           participantRefs,
           simParams,
-          environmentRefs.scheduler,
+          gridConf.environmentRefs.scheduler,
           maybeControllingEm,
         )
       case input: LoadInput =>
         buildParticipant(
           SimpleInputContainer(input),
-          participantConfigUtil.getOrDefault[LoadRuntimeConfig](
+          gridConf.participantConfigUtil.getOrDefault[LoadRuntimeConfig](
             input.getUuid
           ),
-          outputConfigUtil.getOrDefault(NotifierIdentifier.Load),
+          gridConf.outputConfigUtil.getOrDefault(NotifierIdentifier.Load),
           participantRefs,
           simParams,
-          environmentRefs.scheduler,
+          gridConf.environmentRefs.scheduler,
           maybeControllingEm,
         )
       case input: PvInput =>
         buildParticipant(
           SimpleInputContainer(input),
-          participantConfigUtil.getOrDefault[PvRuntimeConfig](
+          gridConf.participantConfigUtil.getOrDefault[PvRuntimeConfig](
             input.getUuid
           ),
-          outputConfigUtil.getOrDefault(NotifierIdentifier.PvPlant),
+          gridConf.outputConfigUtil.getOrDefault(NotifierIdentifier.PvPlant),
           participantRefs,
           simParams,
-          environmentRefs.scheduler,
+          gridConf.environmentRefs.scheduler,
           maybeControllingEm,
         )
       case input: BmInput =>
         buildParticipant(
           SimpleInputContainer(input),
-          participantConfigUtil.getOrDefault[BmRuntimeConfig](
+          gridConf.participantConfigUtil.getOrDefault[BmRuntimeConfig](
             input.getUuid
           ),
-          outputConfigUtil.getOrDefault(NotifierIdentifier.BioMassPlant),
+          gridConf.outputConfigUtil.getOrDefault(
+            NotifierIdentifier.BioMassPlant
+          ),
           participantRefs,
           simParams,
-          environmentRefs.scheduler,
+          gridConf.environmentRefs.scheduler,
           maybeControllingEm,
         )
       case input: WecInput =>
         buildParticipant(
           SimpleInputContainer(input),
-          participantConfigUtil.getOrDefault[WecRuntimeConfig](
+          gridConf.participantConfigUtil.getOrDefault[WecRuntimeConfig](
             input.getUuid
           ),
-          outputConfigUtil.getOrDefault(NotifierIdentifier.Wec),
+          gridConf.outputConfigUtil.getOrDefault(NotifierIdentifier.Wec),
           participantRefs,
           simParams,
-          environmentRefs.scheduler,
+          gridConf.environmentRefs.scheduler,
           maybeControllingEm,
         )
       case input: EvcsInput =>
         buildParticipant(
           SimpleInputContainer(input),
-          participantConfigUtil.getOrDefault[EvcsRuntimeConfig](
+          gridConf.participantConfigUtil.getOrDefault[EvcsRuntimeConfig](
             input.getUuid
           ),
-          outputConfigUtil.getOrDefault(NotifierIdentifier.Evcs),
+          gridConf.outputConfigUtil.getOrDefault(NotifierIdentifier.Evcs),
           participantRefs,
           simParams,
-          environmentRefs.scheduler,
+          gridConf.environmentRefs.scheduler,
           maybeControllingEm,
         )
       case input: HpInput =>
@@ -430,13 +389,13 @@ class GridAgentBuilder(
           case Some(thermalGrid) =>
             buildParticipant(
               WithHeatInputContainer(input, thermalGrid),
-              participantConfigUtil.getOrDefault[HpRuntimeConfig](
+              gridConf.participantConfigUtil.getOrDefault[HpRuntimeConfig](
                 input.getUuid
               ),
-              outputConfigUtil.getOrDefault(NotifierIdentifier.Hp),
+              gridConf.outputConfigUtil.getOrDefault(NotifierIdentifier.Hp),
               participantRefs,
               simParams,
-              environmentRefs.scheduler,
+              gridConf.environmentRefs.scheduler,
               maybeControllingEm,
             )
           case None =>
@@ -447,13 +406,13 @@ class GridAgentBuilder(
       case input: StorageInput =>
         buildParticipant(
           SimpleInputContainer(input),
-          participantConfigUtil.getOrDefault[StorageRuntimeConfig](
+          gridConf.participantConfigUtil.getOrDefault[StorageRuntimeConfig](
             input.getUuid
           ),
-          outputConfigUtil.getOrDefault(NotifierIdentifier.Storage),
+          gridConf.outputConfigUtil.getOrDefault(NotifierIdentifier.Storage),
           participantRefs,
           simParams,
-          environmentRefs.scheduler,
+          gridConf.environmentRefs.scheduler,
           maybeControllingEm,
         )
       case input: SystemParticipantInput =>
@@ -471,6 +430,8 @@ class GridAgentBuilder(
       simParams: SimulationParameters,
       scheduler: ActorRef[SchedulerMessage],
       maybeControllingEm: Option[ActorRef[FlexResponse]],
+  )(implicit
+      gridAgentContext: ActorContext[GridAgent.Request]
   ): ActorRef[ParticipantAgent.Request] = {
 
     val key = ScheduleLock.singleKey(gridAgentContext, scheduler, PRE_INIT_TICK)
@@ -500,10 +461,6 @@ class GridAgentBuilder(
     *
     * @param emInput
     *   The input model
-    * @param modelConfiguration
-    *   Runtime configuration for the agent
-    * @param outputConfig
-    *   Configuration for output notification
     * @param maybeControllingEm
     *   The parent EmAgent, if applicable
     * @return
@@ -511,21 +468,22 @@ class GridAgentBuilder(
     */
   private def buildEm(
       emInput: EmInput,
-      modelConfiguration: EmRuntimeConfig,
-      outputConfig: NotifierConfig,
       maybeControllingEm: Option[ActorRef[FlexResponse]],
+  )(implicit
+      gridConf: GridAgentConstantData,
+      gridAgentContext: ActorContext[GridAgent.Request],
   ): ActorRef[FlexResponse] =
     gridAgentContext.spawn(
       EmAgent(
         emInput,
-        modelConfiguration,
-        outputConfig,
+        gridConf.emConfigUtil.getOrDefault(emInput.getUuid),
+        gridConf.outputConfigUtil.getOrDefault(NotifierIdentifier.Em),
         emInput.getControlStrategy,
-        simulationStartDate,
+        gridConf.simStartTime,
         maybeControllingEm.toRight(
-          environmentRefs.scheduler
+          gridConf.environmentRefs.scheduler
         ),
-        listener,
+        gridConf.listener,
       ),
       actorName(classOf[EmAgent.type], emInput.getId),
     )